+++ client/src/game/ui/selection/ShapeContext.vue
