+++ client/src/game/ui/tools/SelectTool.vue
@@ -9,7 +9,7 @@ import { snapToPoint } from "@/game/layers/utils";
 import { Shape } from "@/game/shapes/shape";
 import { Rect } from "@/game/shapes/variants/rect";
 import { DEFAULT_GRID_SIZE, gameStore } from "@/game/store";
-import Tool from "@/game/ui/tools/dTool.vue";
+import Tool from "@/game/ui/tools/Tool.vue";
 import { calculateDelta, ToolFeatures, ToolName, ToolPermission } from "@/game/ui/tools/utils";
 import { g2l, g2lx, g2ly, l2g, l2gz } from "@/game/units";
 import { equalPoints, getLocalPointFromEvent, useSnapping } from "@/game/utils";
@@ -31,7 +31,7 @@ import { Line } from "../../shapes/variants/line";
 
 import { RulerFeatures } from "./RulerTool.vue";
 import { ToolBasics } from "./ToolBasics";
-import Tools from "./dTools.vue";
+import Tools from "./Tools.vue";
 
 enum SelectOperations {
     Noop,
